move models into a package
