+++ server/models/__init__.py
@@ -1,0 +1,5 @@
+from .db import db
+from .team import Team
+from .player import Player
+from .game import Game
+from .stats import Stats
