+++ server/models/stats.py
@@ -1,75 +1,5 @@
-from flask_sqlalchemy import SQLAlchemy
+from .db import db
 import json
-
-db = SQLAlchemy()
-
-class Team(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.Text)
-
-
-class Player(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    zuluru_id = db.Column(db.Integer, unique=True)
-    team_id = db.Column(db.Integer, db.ForeignKey('team.id'))
-    name = db.Column(db.Text)
-    gender = db.Column(db.Text)
-
-    @property
-    def is_male(self):
-        return self.gender == 'male'
-
-    @property
-    def team(self):
-        return Team.query.get(self.team_id)
-
-    @property
-    def salary(self):
-        pro_rated_number_of_points = 15
-        pro_rated_salary = self._avg_salary_per_point_based_on_history * pro_rated_number_of_points
-        return round(pro_rated_salary)
-
-    @property
-    def _avg_salary_per_point_based_on_history(self):
-        player_stats = Stats.query.filter_by(player_id=self.id)
-        salaries = [ps.salary_per_point for ps in player_stats]
-
-        if len(salaries) > 0:
-            return sum(salaries) / len(salaries)
-        else:
-            return 0
-
-
-class Game(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    league = db.Column(db.Text)
-    week = db.Column(db.Integer)
-    home_team = db.Column(db.Text)
-    away_team = db.Column(db.Text)
-    home_roster = db.Column(db.Text)
-    away_roster = db.Column(db.Text)
-    home_score = db.Column(db.Integer)
-    away_score = db.Column(db.Integer)
-    points = db.Column(db.Text)
-
-    @property
-    def players(self):
-        return json.loads(self.home_roster) + json.loads(self.away_roster)
-
-    def to_dict(self):
-        return {
-            "id": self.id,
-            "league": self.league,
-            "week": self.week,
-            "homeTeam": self.home_team,
-            "homeScore": self.home_score,
-            "homeRoster": json.loads(self.home_roster),
-            "awayTeam": self.away_team,
-            "awayScore": self.away_score,
-            "awayRoster": json.loads(self.away_roster),
-            "points": json.loads(self.points)
-        }
-
 
 class Stats(db.Model):
     SALARY = {
