+++ server/models/team.py
@@ -1,0 +1,6 @@
+from .db import db
+import json
+
+class Team(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.Text)
